src/app/api/pyproject.toml
[project]
name = "app"
version = "0.1.0"
readme = "README.md"
requires-python = "==3.13.*"
# only include packages that are actually imported or required at runtime
dependencies = [
    "asyncpg>=0.30.0",
    "fastapi==0.115.4",
    "pydantic-settings>=2.10.1",
    "sqlalchemy>=2.0.43",
    "pyjwt>=2.10.1",
]


[dependency-groups]
dev = [
    # development and lint/test tooling
    "fastapi[standard]",
    "pytest",
    "pytest-cov>=4.0",   # provide pytest-cov so `--cov` args work in CI
    "mypy>=1.10",        # type checker
    "ruff>=0.14.2",
    "pyinstrument>=5.1.1",
    # keep psycopg as an optional helper for CI but not runtime
    "psycopg>=3.2",
]

[build-system]
requires = ["uv_build"]
build-backend = "uv_build"

[tool.ruff]
src = ["src" , "tests"]
line-length = 150
indent-width = 4

[tool.ruff.format]
quote-style = "double"
indent-style = "space"
skip-magic-trailing-comma = false
line-ending = "auto"

[tool.ruff.lint]
select = ["E4", "E7", "E9", "F"]
